Add support for kafka 2.1 client
