+++ nifi-nar-bundles/nifi-kafka-bundle/nifi-kafka-2-1-processors/src/main/java/org/apache/nifi/record/sink/kafka/KafkaRecordSink_2_1.java
@@ -70,7 +70,7 @@ import java.util.concurrent.TimeoutException;
 
 @Tags({"kafka", "record", "sink"})
 @CapabilityDescription("Provides a service to write records to a Kafka 2.x topic.")
-public class KafkaRecordSink_2_0 extends AbstractControllerService implements RecordSinkService {
+public class KafkaRecordSink_2_1 extends AbstractControllerService implements RecordSinkService {
 
     static final AllowableValue DELIVERY_REPLICATED = new AllowableValue("all", "Guarantee Replicated Delivery",
             "Records are considered 'transmitted unsuccessfully' unless the message is replicated to the appropriate "
@@ -143,7 +143,7 @@ public class KafkaRecordSink_2_0 extends AbstractControllerService implements Re
             .description("This parameter allows you to specify the compression codec for all data generated by this producer.")
             .required(true)
             .addValidator(StandardValidators.NON_EMPTY_VALIDATOR)
-            .allowableValues("none", "gzip", "snappy", "lz4")
+            .allowableValues("none", "gzip", "snappy", "lz4", "zstd")
             .defaultValue("none")
             .build();
