+++ nifi-nar-bundles/nifi-kafka-bundle/nifi-kafka-2-1-processors/src/main/resources/META-INF/services/org.apache.nifi.controller.ControllerService
@@ -12,4 +12,4 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-org.apache.nifi.record.sink.kafka.KafkaRecordSink_2_0
+org.apache.nifi.record.sink.kafka.KafkaRecordSink_2_1
