+++ nifi-nar-bundles/nifi-kafka-bundle/nifi-kafka-2-1-processors/src/main/resources/META-INF/services/org.apache.nifi.processor.Processor
@@ -12,7 +12,7 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-org.apache.nifi.processors.kafka.pubsub.PublishKafka_2_0
-org.apache.nifi.processors.kafka.pubsub.PublishKafkaRecord_2_0
-org.apache.nifi.processors.kafka.pubsub.ConsumeKafka_2_0
-org.apache.nifi.processors.kafka.pubsub.ConsumeKafkaRecord_2_0
+org.apache.nifi.processors.kafka.pubsub.PublishKafka_2_1
+org.apache.nifi.processors.kafka.pubsub.PublishKafkaRecord_2_1
+org.apache.nifi.processors.kafka.pubsub.ConsumeKafka_2_1
+org.apache.nifi.processors.kafka.pubsub.ConsumeKafkaRecord_2_1
