+++ nifi-nar-bundles/nifi-kafka-bundle/nifi-kafka-2-1-processors/src/main/resources/docs/org.apache.nifi.processors.kafka.pubsub.ConsumeKafkaRecord_2_1/additionalDetails.html
